+++ cjs/App-Bo2cpotc.js
@@ -1,32 +1,58 @@
-import { createBrowserHistory, createMemoryHistory } from 'history';
-import { takeEvery, select, put, call, all } from '@redux-saga/core/effects';
-import * as log from 'loglevel';
-import { G as GET_NODE_TREE, b as SET_NODE_TREE, d as GET_NODE_TREE_ERROR } from './store-BvW6_C0H.js';
-import { h as hasNavigationTree, b as injectRedux } from './version-DIHIOZQD.js';
-import to, { to as to$1 } from 'await-to-js';
-import { Query, Op } from 'contensis-delivery-api';
-import { a as actions } from './slice-C6JLQik8.js';
-import { o as selectCurrentPath, G as setRoute, K as selectRouteEntryAvailableLanguages, M as selectRouteEntryID, N as selectStaticRoute, y as SET_NAVIGATION_PATH, x as SET_ROUTE, j as selectRouteEntry, O as selectCurrentNode, f as selectCurrentProject, z as UPDATE_LOADING_STATE, P as selectCurrentAncestors, Q as selectCurrentSiblings, T as selectRouteEntryEntryId, W as selectRouteEntryLanguage, n as selectMappedEntry, A as SET_ENTRY, B as SET_ANCESTORS, w as SET_SIBLINGS, R as REGISTER_USER, X as REGISTER_USER_SUCCESS, Y as REGISTER_USER_FAILED, E as queryParams, F as selectCurrentSearch, H as REQUEST_USER_PASSWORD_RESET, I as RESET_USER_PASSWORD, J as CHANGE_USER_PASSWORD, Z as REQUEST_USER_PASSWORD_RESET_SENDING, _ as REQUEST_USER_PASSWORD_RESET_SUCCESS, $ as REQUEST_USER_PASSWORD_RESET_ERROR, a0 as RESET_USER_PASSWORD_SENDING, a1 as RESET_USER_PASSWORD_SUCCESS, a2 as RESET_USER_PASSWORD_ERROR, a3 as CHANGE_USER_PASSWORD_ERROR, a4 as CHANGE_USER_PASSWORD_SENDING, a5 as CHANGE_USER_PASSWORD_SUCCESS } from './selectors-PJo8AWy0.js';
-import { s as selectVersionStatus } from './version-DlaBPQ7d.js';
-import { a as cachedSearch } from './ContensisDeliveryApi-CqPcnP_R.js';
-import React from 'react';
-import 'react-cookie';
-import 'react-redux';
-import './VersionInfo-BMAAda1K.js';
-import 'jsonpath-mapper';
-import 'react-router-dom';
-import { s as selectCurrentLanguage, a as selectDictionary, c as selectLocaleRoutes, b as selectLocales, f as selectDictionaryResolver, d as selectPrimaryLanguage } from './selectors-DcmvOeX2.js';
-import { h as handleRequiresLoginSaga, L as LoginHelper, i as findContentTypeMapping, j as getSearchOptions, k as getManagementApiClient, m as loginSagas } from './ChangePassword.container-giznBLAf.js';
-import './sagas-CFEhiMQT.js';
-import 'reselect';
-import 'immer';
-import 'deep-equal';
-import 'deepmerge';
-import { r as routeParams } from './util-BuAZdfcq.js';
-import 'contensis-core-api';
-import { l as selectClientCredentials } from './matchGroups-DT-RunAc.js';
-import { a as RouteLoader } from './RouteLoader-yDf3uwGr.js';
-import 'query-string';
+'use strict';
+
+var history$1 = require('history');
+var effects = require('@redux-saga/core/effects');
+var log = require('loglevel');
+var store = require('./store-DxuhGQ7p.js');
+var version = require('./version-B9moFk8k.js');
+var to = require('await-to-js');
+var contensisDeliveryApi = require('contensis-delivery-api');
+var slice = require('./slice-5xJMH24n.js');
+var selectors$1 = require('./selectors-C1CqEUmL.js');
+var version$1 = require('./version-CukCz8zL.js');
+var ContensisDeliveryApi = require('./ContensisDeliveryApi-DBaziPG9.js');
+var React = require('react');
+require('react-cookie');
+require('react-redux');
+require('./VersionInfo-CTPtw_Xd.js');
+require('jsonpath-mapper');
+require('react-router-dom');
+var selectors = require('./selectors-DAQR0uZa.js');
+var ChangePassword_container = require('./ChangePassword.container-BWh4R32r.js');
+require('./sagas-C6YSqau4.js');
+require('reselect');
+require('immer');
+require('deep-equal');
+require('deepmerge');
+var util = require('./util-CFAjDCA0.js');
+require('contensis-core-api');
+var matchGroups = require('./matchGroups-CxRa9Ej9.js');
+var RouteLoader = require('./RouteLoader-B-sjh6ha.js');
+require('query-string');
+
+function _interopDefault (e) { return e && e.__esModule ? e : { default: e }; }
+
+function _interopNamespace(e) {
+  if (e && e.__esModule) return e;
+  var n = Object.create(null);
+  if (e) {
+    Object.keys(e).forEach(function (k) {
+      if (k !== 'default') {
+        var d = Object.getOwnPropertyDescriptor(e, k);
+        Object.defineProperty(n, k, d.get ? d : {
+          enumerable: true,
+          get: function () { return e[k]; }
+        });
+      }
+    });
+  }
+  n.default = e;
+  return Object.freeze(n);
+}
+
+var log__namespace = /*#__PURE__*/_interopNamespace(log);
+var to__default = /*#__PURE__*/_interopDefault(to);
+var React__default = /*#__PURE__*/_interopDefault(React);
 
 const deparameterise = path => {
   return path.replace(/\/:\w+\??/g, '');
@@ -68,11 +94,11 @@ const createLocaleRoutes = routes => {
 };
 
 // Create a history depending on the environment
-const selectedHistory = typeof window !== 'undefined' ? createBrowserHistory : createMemoryHistory;
+const selectedHistory = typeof window !== 'undefined' ? history$1.createBrowserHistory : history$1.createMemoryHistory;
 const history = (options = {}) => selectedHistory(options);
 const browserHistory = selectedHistory();
 
-const navigationSagas = [takeEvery(GET_NODE_TREE, ensureNodeTreeSaga)];
+const navigationSagas = [effects.takeEvery(store.GET_NODE_TREE, ensureNodeTreeSaga)];
 function* ensureNodeTreeSaga(action) {
   const {
     api,
@@ -81,34 +107,34 @@ function* ensureNodeTreeSaga(action) {
     versionStatus,
     treeDepth
   } = action;
-  const state = yield select();
+  const state = yield effects.select();
   try {
-    if (!hasNavigationTree(state)) {
+    if (!version.hasNavigationTree(state)) {
       const nodes = yield api.getRootNode({
         depth: treeDepth || 0,
         language
       }, versionStatus, project);
       if (nodes) {
-        yield put({
-          type: SET_NODE_TREE,
+        yield effects.put({
+          type: store.SET_NODE_TREE,
           nodes
         });
       } else {
-        yield put({
-          type: GET_NODE_TREE_ERROR
+        yield effects.put({
+          type: store.GET_NODE_TREE_ERROR
         });
       }
     }
   } catch (ex) {
-    log.error(...['Error running ensureNodeTreeSaga:', ex]);
-    yield put({
-      type: GET_NODE_TREE_ERROR,
+    log__namespace.error(...['Error running ensureNodeTreeSaga:', ex]);
+    yield effects.put({
+      type: store.GET_NODE_TREE_ERROR,
       error: ex.toString()
     });
   }
 }
 
-const i18nSagas = [takeEvery(actions.INIT_LOCALES.type, getProjectLanguages), takeEvery(actions.UPDATE_LANGUAGE.type, updateLanguage), takeEvery(actions.SET_LANGUAGE.type, setLanguageRoute)];
+const i18nSagas = [effects.takeEvery(slice.actions.INIT_LOCALES.type, getProjectLanguages), effects.takeEvery(slice.actions.UPDATE_LANGUAGE.type, updateLanguage), effects.takeEvery(slice.actions.SET_LANGUAGE.type, setLanguageRoute)];
 
 /**
  * Resolve the current route language based on the entry, node, static route or path
@@ -119,26 +145,26 @@ function* resolveCurrentRouteLanguage({
   node
 }) {
   var _entry$sys, _staticRoute$route;
-  const currentLanguage = yield select(selectCurrentLanguage);
-  const staticRoute = yield select(selectStaticRoute);
+  const currentLanguage = yield effects.select(selectors.selectCurrentLanguage);
+  const staticRoute = yield effects.select(selectors$1.selectStaticRoute);
   let nextLanguage = currentLanguage;
   if (entry !== null && entry !== void 0 && (_entry$sys = entry.sys) !== null && _entry$sys !== void 0 && _entry$sys.language) nextLanguage = entry.sys.language;else if (node !== null && node !== void 0 && node.language) nextLanguage = node.language;else if (staticRoute !== null && staticRoute !== void 0 && (_staticRoute$route = staticRoute.route) !== null && _staticRoute$route !== void 0 && _staticRoute$route.language) nextLanguage = staticRoute.route.language;else {
     var _currentPath$split$fi;
     // attempt to infer language from the path
-    const currentPath = yield select(selectCurrentPath);
+    const currentPath = yield effects.select(selectors$1.selectCurrentPath);
 
     // path is normally lowercase
     const firstPathSegment = (_currentPath$split$fi = currentPath.split('/').find(segment => segment.length)) === null || _currentPath$split$fi === void 0 ? void 0 : _currentPath$split$fi.toLowerCase();
-    const locales = yield select(selectLocales);
+    const locales = yield effects.select(selectors.selectLocales);
     const matchedLanguage = Object.keys(locales).find(lang => lang.toLowerCase() === firstPathSegment);
     // matched a supported language in the path
     if (matchedLanguage) nextLanguage = matchedLanguage;else
       // falling back to primary language
-      nextLanguage = yield select(selectPrimaryLanguage);
+      nextLanguage = yield effects.select(selectors.selectPrimaryLanguage);
   }
   if (nextLanguage && nextLanguage !== currentLanguage) {
-    const dictionary = yield call(resolveDictionaryForLanguage, nextLanguage);
-    yield put(actions.SET_LANGUAGE({
+    const dictionary = yield effects.call(resolveDictionaryForLanguage, nextLanguage);
+    yield effects.put(slice.actions.SET_LANGUAGE({
       language: nextLanguage,
       dictionary
     }));
@@ -151,20 +177,20 @@ function* resolveCurrentRouteLanguage({
  * Is called directly any time the language is changed
  */
 function* resolveDictionaryForLanguage(language) {
-  let dictionary = yield select(selectDictionary);
+  let dictionary = yield effects.select(selectors.selectDictionary);
   // try and resolve a dictionary for this language
-  const resolver = yield select(selectDictionaryResolver);
+  const resolver = yield effects.select(selectors.selectDictionaryResolver);
   if (typeof resolver === 'function') {
     try {
       // dynamic import of dictionary file
-      const loadedDictionary = yield call(resolver, language);
+      const loadedDictionary = yield effects.call(resolver, language);
       dictionary = loadedDictionary;
     } catch (error) {
       console.error(`No dictionary resolved for language ${language}`, error);
     }
   } else {
     // Load dictionary from locales in state
-    const locales = yield select(selectLocales);
+    const locales = yield effects.select(selectors.selectLocales);
     if (locales && locales[language]) {
       dictionary = locales[language];
     }
@@ -184,18 +210,18 @@ function* updateLanguage({
     fallbackPath
   }
 }) {
-  const currentLanguage = yield select(selectCurrentLanguage);
+  const currentLanguage = yield effects.select(selectors.selectCurrentLanguage);
   if (language === currentLanguage) {
     // no change needed
     return;
   } else {
-    const dictionary = language !== currentLanguage ? yield call(resolveDictionaryForLanguage, language) : yield select(selectDictionary);
-    const uri = yield call(resolveNextLanguageRoute, {
+    const dictionary = language !== currentLanguage ? yield effects.call(resolveDictionaryForLanguage, language) : yield effects.select(selectors.selectDictionary);
+    const uri = yield effects.call(resolveNextLanguageRoute, {
       language,
       redirect,
       fallbackPath
     });
-    yield put(actions.SET_LANGUAGE({
+    yield effects.put(slice.actions.SET_LANGUAGE({
       language,
       dictionary: dictionary != null ? dictionary : undefined,
       redirect: redirect !== false ? uri : undefined
@@ -208,12 +234,12 @@ function* setLanguageRoute({
   payload
 }) {
   if (payload !== null && payload !== void 0 && payload.redirect) {
-    const currentPath = yield select(selectCurrentPath);
+    const currentPath = yield effects.select(selectors$1.selectCurrentPath);
     if (payload.redirect === currentPath) {
       // already on the correct path, no need to redirect
       return;
     }
-    yield put(setRoute(payload.redirect));
+    yield effects.put(selectors$1.setRoute(payload.redirect));
   }
 }
 
@@ -229,11 +255,11 @@ function* resolveNextLanguageRoute({
   }
 
   // is this an entry or a static route?
-  const availableLanguages = yield select(selectRouteEntryAvailableLanguages);
+  const availableLanguages = yield effects.select(selectors$1.selectRouteEntryAvailableLanguages);
   if (availableLanguages.find(l => l.toLowerCase() === language.toLowerCase())) {
     // if entry, get the uri for this language variation from the api
-    const entryUri = yield call(getEntryUriForLanguage, {
-      entryId: yield select(selectRouteEntryID),
+    const entryUri = yield effects.call(getEntryUriForLanguage, {
+      entryId: yield effects.select(selectors$1.selectRouteEntryID),
       language
     });
     if (entryUri) {
@@ -242,7 +268,7 @@ function* resolveNextLanguageRoute({
   }
 
   // if static route, get the uri from the routes config
-  const staticRouteUri = yield call(getStaticRouteUri, {
+  const staticRouteUri = yield effects.call(getStaticRouteUri, {
     language
   });
   if (staticRouteUri) {
@@ -257,13 +283,13 @@ function* resolveNextLanguageRoute({
 function* getStaticRouteUri({
   language
 }) {
-  const staticRoute = yield select(selectStaticRoute);
+  const staticRoute = yield effects.select(selectors$1.selectStaticRoute);
   if (staticRoute !== null && staticRoute !== void 0 && staticRoute.route.path) {
     var _Object$entries$find;
     // Routes can have parameters such as `/:facet?` we need to deparameterise
     // so we can check against our stored locale routes
     const deparameterisedPath = deparameterise(staticRoute.route.path);
-    const localeRoutes = yield select(selectLocaleRoutes);
+    const localeRoutes = yield effects.select(selectors.selectLocaleRoutes);
     const originalPath = (_Object$entries$find = Object.entries(localeRoutes || {}).find(([, locales]) => Object.values(locales).includes(deparameterisedPath))) === null || _Object$entries$find === void 0 ? void 0 : _Object$entries$find[0];
     const routeLocales = localeRoutes[deparameterisedPath] || localeRoutes[originalPath || ''];
     const routeUri = routeLocales === null || routeLocales === void 0 ? void 0 : routeLocales[language];
@@ -278,7 +304,7 @@ function* getStaticRouteUri({
 function* getProjectLanguages({
   payload
 }) {
-  const stateLocales = yield select(selectLocales);
+  const stateLocales = yield effects.select(selectors.selectLocales);
   if (stateLocales && Object.keys(stateLocales).length > 0)
     // Locales already set in state, no need to fetch again
     return;
@@ -292,7 +318,7 @@ function* getProjectLanguages({
     }
   } else {
     // Fallback to getting languages from the project
-    const [error, project] = yield to(cachedSearch.getClient().project.get());
+    const [error, project] = yield to__default.default(ContensisDeliveryApi.cachedSearch.getClient().project.get());
     if (error) {
       console.error('Problem fetching project languages:', error);
     } else if (project) {
@@ -312,7 +338,7 @@ function* getProjectLanguages({
   }
 
   // Only commit if we have locales to set or we will end up in an infinite loop
-  if (Object.keys(locales).length) yield put(actions.SET_LOCALES({
+  if (Object.keys(locales).length) yield effects.put(slice.actions.SET_LOCALES({
     ...payload,
     primaryLanguage,
     supportedLanguages,
@@ -328,15 +354,15 @@ function* getEntryUriForLanguage({
   language
 }) {
   try {
-    const versionStatus = yield select(selectVersionStatus);
-    const query = new Query(Op.equalTo('sys.id', entryId), Op.equalTo('sys.language', language), Op.equalTo('sys.versionStatus', versionStatus));
+    const versionStatus = yield effects.select(version$1.selectVersionStatus);
+    const query = new contensisDeliveryApi.Query(contensisDeliveryApi.Op.equalTo('sys.id', entryId), contensisDeliveryApi.Op.equalTo('sys.language', language), contensisDeliveryApi.Op.equalTo('sys.versionStatus', versionStatus));
     query.fields = ['sys.uri'];
     query.pageSize = 1;
-    const result = yield cachedSearch.search(query);
+    const result = yield ContensisDeliveryApi.cachedSearch.search(query);
     return result.items.length ? result.items[0].sys.uri : null;
   } catch (error) {
     console.error('Error fetching language variations:', error);
-    yield put(actions.GET_ENTRY_URI_ERROR(error));
+    yield effects.put(slice.actions.GET_ENTRY_URI_ERROR(error));
   }
 }
 
@@ -348,15 +374,15 @@ const Fields = {
 
 const fieldExpression = (field, value, operator = 'equalTo', weight = null) => {
   if (!field || !value) return [];
-  if (Array.isArray(value)) return equalToOrIn(field, value, operator);else return !weight ? [Op[operator](field, value)] : [Op[operator](field, value).weight(weight)];
+  if (Array.isArray(value)) return equalToOrIn(field, value, operator);else return !weight ? [contensisDeliveryApi.Op[operator](field, value)] : [contensisDeliveryApi.Op[operator](field, value).weight(weight)];
 };
 const defaultExpressions = versionStatus => {
-  return [Op.equalTo(Fields.sys.versionStatus, versionStatus)];
+  return [contensisDeliveryApi.Op.equalTo(Fields.sys.versionStatus, versionStatus)];
 };
-const equalToOrIn = (field, arr, operator = 'equalTo') => arr.length === 0 ? [] : arr.length === 1 ? [Op[operator](field, arr[0])] : [Op.in(field, ...arr)];
+const equalToOrIn = (field, arr, operator = 'equalTo') => arr.length === 0 ? [] : arr.length === 1 ? [contensisDeliveryApi.Op[operator](field, arr[0])] : [contensisDeliveryApi.Op.in(field, ...arr)];
 
 const routeEntryByFieldsQuery = (id, language = 'en-GB', contentTypeId = '', fields = [], fieldLinkDepths, versionStatus = 'published') => {
-  const query = new Query(...[...fieldExpression('sys.id', id), ...fieldExpression('sys.language', language), ...(contentTypeId ? fieldExpression('sys.contentTypeId', contentTypeId) : fieldExpression('sys.dataFormat', 'entry')), ...defaultExpressions(versionStatus)]);
+  const query = new contensisDeliveryApi.Query(...[...fieldExpression('sys.id', id), ...fieldExpression('sys.language', language), ...(contentTypeId ? fieldExpression('sys.contentTypeId', contentTypeId) : fieldExpression('sys.dataFormat', 'entry')), ...defaultExpressions(versionStatus)]);
   query.fields = fields;
   query.fieldLinkDepths = fieldLinkDepths;
   return query;
@@ -369,7 +395,7 @@ function* reduxInjectorSaga(injectorFn) {
       reducer,
       saga
     } = yield injectorFn();
-    injectRedux({
+    version.injectRedux({
       key,
       reducer,
       saga
@@ -380,7 +406,7 @@ function* reduxInjectorSaga(injectorFn) {
 /**
  * @description Asynchronously load and inject assets related to Search
  */
-const importSearchAssets = () => import('./search.js');
+const importSearchAssets = () => Promise.resolve().then(function () { return require('./search.js'); });
 
 /**
  * Invokes the Search saga if:
@@ -428,12 +454,12 @@ function* handleSearchSaga({
     } = yield importSearchAssets();
 
     // Inject search reducer and sagas
-    yield call(reduxInjectorSaga, async () => ({
+    yield effects.call(reduxInjectorSaga, async () => ({
       key: 'search',
       reducer: searchOpts.config && reducer(searchOpts.config),
       saga: sagas
     }));
-    if (invokeSearch) yield call(setRouteFilters, {
+    if (invokeSearch) yield effects.call(setRouteFilters, {
       params,
       ssr,
       ...searchOpts,
@@ -442,14 +468,14 @@ function* handleSearchSaga({
   }
 }
 
-const routingSagas = [takeEvery(SET_NAVIGATION_PATH, getRouteSaga), takeEvery(SET_ROUTE, setRouteSaga)];
+const routingSagas = [effects.takeEvery(selectors$1.SET_NAVIGATION_PATH, getRouteSaga), effects.takeEvery(selectors$1.SET_ROUTE, setRouteSaga)];
 
 /**
  * To navigate / push a specific route via redux middleware
  * @param {path, state} action
  */
 function* setRouteSaga(action) {
-  yield put({
+  yield effects.put({
     type: 'CALL_HISTORY_METHOD',
     payload: {
       method: 'push',
@@ -474,7 +500,7 @@ function* getRouteSaga(action) {
     } = action;
 
     // Inject redux { key, reducer, saga } provided by staticRoute
-    if (staticRoute && staticRoute.route.injectRedux) yield call(reduxInjectorSaga, staticRoute.route.injectRedux);
+    if (staticRoute && staticRoute.route.injectRedux) yield effects.call(reduxInjectorSaga, staticRoute.route.injectRedux);
 
     // Variables we will pass to setRouteEntry
     let pathNode = null,
@@ -504,15 +530,15 @@ function* getRouteSaga(action) {
     const entryFieldLinkDepths = (_appsays = appsays) === null || _appsays === void 0 ? void 0 : _appsays.entryFieldLinkDepths;
     const setStaticRouteLimits = typeof linkDepth !== 'undefined' || fields || fieldLinkDepths;
     const setContentTypeLimits = !!ContentTypeMappings.find(ct => ct.fields || ct.linkDepth || ct.nodeOptions || ct.fieldLinkDepths);
-    const state = yield select();
-    const routeEntry = selectRouteEntry(state, 'js');
-    const routeNode = selectCurrentNode(state, 'js');
+    const state = yield effects.select();
+    const routeEntry = selectors$1.selectRouteEntry(state, 'js');
+    const routeNode = selectors$1.selectCurrentNode(state, 'js');
     const currentPath = action.path; //selectCurrentPath(state);
-    const deliveryApiStatus = selectVersionStatus(state);
-    const project = selectCurrentProject(state);
+    const deliveryApiStatus = version$1.selectVersionStatus(state);
+    const project = selectors$1.selectCurrentProject(state);
     // const isHome = currentPath === '/';
     const isPreview = currentPath && currentPath.startsWith('/preview/');
-    const currentLanguage = selectCurrentLanguage(state);
+    const currentLanguage = selectors.selectCurrentLanguage(state);
     const defaultLang = appsays && appsays.defaultLang || currentLanguage || 'en-GB';
     if (!isPreview && ((_appsays2 = appsays) !== null && _appsays2 !== void 0 && _appsays2.customRouting || staticRoute && !staticRoute.route.fetchNode || routeEntry && action.statePath === action.path && ((_appsays3 = appsays) === null || _appsays3 === void 0 ? void 0 : _appsays3.refetchNode) !== true)) {
       var _staticRoute$route5;
@@ -526,11 +552,11 @@ function* getRouteSaga(action) {
           entry: null
         };
         pathNode.entry = entry = routeEntry;
-        yield put({
-          type: UPDATE_LOADING_STATE,
+        yield effects.put({
+          type: selectors$1.UPDATE_LOADING_STATE,
           isLoading: false
         });
-      } else yield call(setRouteEntry, currentPath, routeEntry, yield select(selectCurrentNode), yield select(selectCurrentAncestors), yield select(selectCurrentSiblings));
+      } else yield effects.call(setRouteEntry, currentPath, routeEntry, yield effects.select(selectors$1.selectCurrentNode), yield effects.select(selectors$1.selectCurrentAncestors), yield effects.select(selectors$1.selectCurrentSiblings));
     } else {
       var _staticRoute$route6;
       // Handle preview routes
@@ -566,7 +592,7 @@ function* getRouteSaga(action) {
         let nodeError = undefined;
         // Resolve a stub of route node if we are setting limits in content type mappings
         // Resolve the complete entry with the node if we are setting limits in a static route
-        [nodeError, pathNode] = yield to(api.getNode({
+        [nodeError, pathNode] = yield to__default.default(api.getNode({
           depth: 0,
           path: currentPath,
           entryFields: setStaticRouteLimits ? fields || '*' : setContentTypeLimits ? ['sys.contentTypeId', 'sys.id'] : '*',
@@ -583,22 +609,22 @@ function* getRouteSaga(action) {
             // and fire the user down the handleRequiresLoginSaga
             // If auth was successful via a refreshToken we need to reload the page
             // to run this getRouteSaga again with the security token cookie
-            const userLoggedIn = yield call(handleRequiresLoginSaga, {
+            const userLoggedIn = yield effects.call(ChangePassword_container.handleRequiresLoginSaga, {
               ...action,
               requireLogin: true
             });
             if (userLoggedIn && nodeError.status === 401) {
               // Reload the route so we can re-run the routing request now the
               // authentication cookies are written
-              return yield call(setRouteSaga, {
+              return yield effects.call(setRouteSaga, {
                 path: currentPath
               });
             } else if (userLoggedIn && nodeError.status === 403) {
-              return yield call(setRouteSaga, {
-                path: LoginHelper.GetAccessDeniedRoute(currentPath)
+              return yield effects.call(setRouteSaga, {
+                path: ChangePassword_container.LoginHelper.GetAccessDeniedRoute(currentPath)
               });
             } else {
-              return yield call(do500, nodeError);
+              return yield effects.call(do500, nodeError);
             }
           } else throw nodeError;
         } else ({
@@ -609,7 +635,7 @@ function* getRouteSaga(action) {
         if ((_pathNode = pathNode) !== null && _pathNode !== void 0 && (_pathNode = _pathNode.entry) !== null && _pathNode !== void 0 && (_pathNode = _pathNode.sys) !== null && _pathNode !== void 0 && _pathNode.id && pathNode.entry.sys.contentTypeId) {
           // Get fields[] and linkDepth from ContentTypeMapping to get the entry data
           // and current node's ordinates at a specified depth with specified fields
-          contentTypeMapping = findContentTypeMapping(ContentTypeMappings, pathNode.entry.sys.contentTypeId);
+          contentTypeMapping = ChangePassword_container.findContentTypeMapping(ContentTypeMappings, pathNode.entry.sys.contentTypeId);
         }
 
         // Run a second search query if we aren't setting limits from a static route
@@ -636,7 +662,7 @@ function* getRouteSaga(action) {
 
       // make calls to fetch node ancestors, children,
       // siblings or entire node tree
-      [ancestors, children, siblings] = yield call(resolveCurrentNodeOrdinates, {
+      [ancestors, children, siblings] = yield effects.call(resolveCurrentNodeOrdinates, {
         api,
         appsays,
         contentTypeMapping: contentTypeMapping || (staticRoute === null || staticRoute === void 0 || (_staticRoute$route6 = staticRoute.route) === null || _staticRoute$route6 === void 0 ? void 0 : _staticRoute$route6.fetchNode) || {},
@@ -652,18 +678,18 @@ function* getRouteSaga(action) {
     // We initially listened for SET_ENTRY to complete before
     // resolving the current route language, but this meant
     // that the language change was not captured in time for the SSR response
-    yield call(resolveCurrentRouteLanguage, {
+    yield effects.call(resolveCurrentRouteLanguage, {
       entry: (_pathNode3 = pathNode) === null || _pathNode3 === void 0 ? void 0 : _pathNode3.entry,
       node: pathNode
     });
-    const contentTypeRoute = findContentTypeMapping(ContentTypeMappings, (_pathNode4 = pathNode) === null || _pathNode4 === void 0 || (_pathNode4 = _pathNode4.entry) === null || _pathNode4 === void 0 || (_pathNode4 = _pathNode4.sys) === null || _pathNode4 === void 0 ? void 0 : _pathNode4.contentTypeId);
+    const contentTypeRoute = ChangePassword_container.findContentTypeMapping(ContentTypeMappings, (_pathNode4 = pathNode) === null || _pathNode4 === void 0 || (_pathNode4 = _pathNode4.entry) === null || _pathNode4 === void 0 || (_pathNode4 = _pathNode4.sys) === null || _pathNode4 === void 0 ? void 0 : _pathNode4.contentTypeId);
 
     // Inject redux { key, reducer, saga } provided by ContentTypeMapping
-    if (contentTypeRoute !== null && contentTypeRoute !== void 0 && contentTypeRoute.injectRedux) yield call(reduxInjectorSaga, contentTypeRoute.injectRedux);
+    if (contentTypeRoute !== null && contentTypeRoute !== void 0 && contentTypeRoute.injectRedux) yield effects.call(reduxInjectorSaga, contentTypeRoute.injectRedux);
 
     // Have we defined search options in the route configuration (for triggering search)
-    const routeSearchOptions = getSearchOptions(staticRoute, contentTypeRoute);
-    const params = routeParams(staticRoute, action.location);
+    const routeSearchOptions = ChangePassword_container.getSearchOptions(staticRoute, contentTypeRoute);
+    const params = util.routeParams(staticRoute, action.location);
     if (withEvents && withEvents.onRouteLoaded) {
       // Check if the app has provided a requireLogin boolean flag or groups array
       // in addition to checking if requireLogin is set in the route definition
@@ -681,13 +707,13 @@ function* getRouteSaga(action) {
     }
     if (requireLogin !== false) {
       // Do not call the login feature saga if requireLogin is false
-      yield call(handleRequiresLoginSaga, {
+      yield effects.call(ChangePassword_container.handleRequiresLoginSaga, {
         ...action,
         entry,
         requireLogin
       });
     }
-    if (searchOptions || routeSearchOptions) yield call(handleSearchSaga, {
+    if (searchOptions || routeSearchOptions) yield effects.call(handleSearchSaga, {
       ...action,
       params,
       routeSearchOptions,
@@ -700,13 +726,13 @@ function* getRouteSaga(action) {
     if ((_pathNode5 = pathNode) !== null && _pathNode5 !== void 0 && (_pathNode5 = _pathNode5.entry) !== null && _pathNode5 !== void 0 && (_pathNode5 = _pathNode5.sys) !== null && _pathNode5 !== void 0 && _pathNode5.id) {
       var _appsays4;
       entry = pathNode.entry;
-      yield call(setRouteEntry, currentPath, entry, pathNode, ancestors, siblings, entryMapper || (contentTypeRoute === null || contentTypeRoute === void 0 ? void 0 : contentTypeRoute.entryMapper), false, (_appsays4 = appsays) === null || _appsays4 === void 0 ? void 0 : _appsays4.refetchNode);
+      yield effects.call(setRouteEntry, currentPath, entry, pathNode, ancestors, siblings, entryMapper || (contentTypeRoute === null || contentTypeRoute === void 0 ? void 0 : contentTypeRoute.entryMapper), false, (_appsays4 = appsays) === null || _appsays4 === void 0 ? void 0 : _appsays4.refetchNode);
     } else {
-      if (staticRoute) yield call(setRouteEntry, currentPath, null, pathNode, ancestors, siblings);else yield call(do404);
+      if (staticRoute) yield effects.call(setRouteEntry, currentPath, null, pathNode, ancestors, siblings);else yield effects.call(do404);
     }
   } catch (e) {
-    log.error(...['Error running route saga:', e, e.stack]);
-    yield call(do500, e);
+    log__namespace.error(...['Error running route saga:', e, e.stack]);
+    yield effects.call(do500, e);
   }
 }
 function* resolveCurrentNodeOrdinates(action) {
@@ -743,7 +769,7 @@ function* resolveCurrentNodeOrdinates(action) {
             versionStatus
           }, project);
         } catch (ex) {
-          log.info('Problem fetching ancestors', ex);
+          log__namespace.info('Problem fetching ancestors', ex);
           return [];
         }
       };
@@ -763,7 +789,7 @@ function* resolveCurrentNodeOrdinates(action) {
             versionStatus
           }, project);
         } catch (ex) {
-          log.info('Problem fetching children', ex);
+          log__namespace.info('Problem fetching children', ex);
           return [];
         }
       };
@@ -782,30 +808,30 @@ function* resolveCurrentNodeOrdinates(action) {
             versionStatus
           }, project);
         } catch (ex) {
-          log.info('Problem fetching siblings', ex);
+          log__namespace.info('Problem fetching siblings', ex);
           return [];
         }
       };
     }
   }
-  const isTreeLoaded = yield select(hasNavigationTree);
+  const isTreeLoaded = yield effects.select(version.hasNavigationTree);
   if (!isTreeLoaded && (doNavigation === true || doNavigation.tree)) apiCall[3] = function* getNodeTree() {
     const treeDepth = doNavigation === true || !doNavigation.tree || doNavigation.tree === true ? 2 : doNavigation.tree;
     if (typeof window !== 'undefined') {
-      return yield put({
-        type: GET_NODE_TREE,
+      return yield effects.put({
+        type: store.GET_NODE_TREE,
         ...action,
         treeDepth
       });
     } else {
-      return yield call(ensureNodeTreeSaga, {
+      return yield effects.call(ensureNodeTreeSaga, {
         ...action,
         treeDepth
       });
     }
   };
   const [loadAncestors, loadChildren, loadSiblings, loadTree] = apiCall;
-  const [ancestors, nodeWithChildren, siblings] = yield all([loadAncestors(), loadChildren(), loadSiblings(), loadTree()]);
+  const [ancestors, nodeWithChildren, siblings] = yield effects.all([loadAncestors(), loadChildren(), loadSiblings(), loadTree()]);
   return [ancestors, nodeWithChildren === null || nodeWithChildren === void 0 ? void 0 : nodeWithChildren.children, siblings];
 }
 function* setRouteEntry(currentPath, entry, node, ancestors, siblings, entryMapper, notFound = false, remapEntry = false) {
@@ -814,54 +840,54 @@ function* setRouteEntry(currentPath, entry, node, ancestors, siblings, entryMapp
   // Update a window global to provide the preview toolbar
   // an updated entry id in client-side navigation
   if (typeof window !== 'undefined') window.ContensisEntryId = entrySys.id;
-  const currentEntryId = yield select(selectRouteEntryEntryId);
-  const currentEntryLang = yield select(selectRouteEntryLanguage);
-  const mappedEntry = !entryMapper ? null : currentEntryId === entrySys.id && currentEntryLang === entrySys.language && remapEntry === false ? (yield select(selectMappedEntry, 'js')) || {} : yield mapRouteEntry(entryMapper, {
+  const currentEntryId = yield effects.select(selectors$1.selectRouteEntryEntryId);
+  const currentEntryLang = yield effects.select(selectors$1.selectRouteEntryLanguage);
+  const mappedEntry = !entryMapper ? null : currentEntryId === entrySys.id && currentEntryLang === entrySys.language && remapEntry === false ? (yield effects.select(selectors$1.selectMappedEntry, 'js')) || {} : yield mapRouteEntry(entryMapper, {
     ...node,
     entry,
     ancestors,
     siblings
   });
-  yield all([put({
-    type: SET_ENTRY,
+  yield effects.all([effects.put({
+    type: selectors$1.SET_ENTRY,
     id: entrySys.id,
     currentPath,
     entry,
     mappedEntry,
     node,
     notFound
-  }), ancestors && put({
-    type: SET_ANCESTORS,
+  }), ancestors && effects.put({
+    type: selectors$1.SET_ANCESTORS,
     ancestors
-  }), siblings && put({
-    type: SET_SIBLINGS,
+  }), siblings && effects.put({
+    type: selectors$1.SET_SIBLINGS,
     siblings
   })]);
 }
 function* mapRouteEntry(entryMapper, node) {
   try {
     if (typeof entryMapper === 'function') {
-      const state = yield select();
-      const mappedEntry = yield call(entryMapper, node, state);
+      const state = yield effects.select();
+      const mappedEntry = yield effects.call(entryMapper, node, state);
       return mappedEntry;
     }
   } catch (e) {
-    log.error(...['Error running entryMapper:', e, e.stack]);
+    log__namespace.error(...['Error running entryMapper:', e, e.stack]);
     throw e;
   }
   return;
 }
 function* do404() {
-  yield call(clientReloadHitServer);
-  yield put({
-    type: SET_ENTRY,
+  yield effects.call(clientReloadHitServer);
+  yield effects.put({
+    type: selectors$1.SET_ENTRY,
     id: null,
     entry: null,
     notFound: true
   });
 }
 function* clientReloadHitServer() {
-  const stateEntry = yield select(selectRouteEntry);
+  const stateEntry = yield effects.select(selectors$1.selectRouteEntry);
 
   // If in client and there is a stateEntry.sys field reload the page,
   // on the 2nd load stateEntry.sys should be null at this point,
@@ -871,8 +897,8 @@ function* clientReloadHitServer() {
   }
 }
 function* do500(error) {
-  yield put({
-    type: SET_ENTRY,
+  yield effects.put({
+    type: selectors$1.SET_ENTRY,
     id: null,
     entry: null,
     notFound: true,
@@ -882,7 +908,7 @@ function* do500(error) {
   });
 }
 
-const registerSagas = [takeEvery(REGISTER_USER, registerSaga), takeEvery(REGISTER_USER_SUCCESS, redirectSaga)];
+const registerSagas = [effects.takeEvery(selectors$1.REGISTER_USER, registerSaga), effects.takeEvery(selectors$1.REGISTER_USER_SUCCESS, redirectSaga)];
 function* registerSaga({
   user,
   mappers
@@ -905,7 +931,7 @@ function* registerSaga({
   });
   if (response.ok) {
     let mappedResponse;
-    const [, responseBody] = yield to$1(response.json());
+    const [, responseBody] = yield to.to(response.json());
     if (responseBody) {
       // Allow use of response mapper to convert the successful user object
       // from the api response body into a user object of any format
@@ -913,14 +939,14 @@ function* registerSaga({
         mappedResponse = yield mappers.response(responseBody);
       }
       // Update user object with mappedResponse or responseBody
-      yield put({
-        type: REGISTER_USER_SUCCESS,
+      yield effects.put({
+        type: selectors$1.REGISTER_USER_SUCCESS,
         user: mappedResponse || responseBody
       });
     } else {
       // OK response but unable to parse the response body
-      yield put({
-        type: REGISTER_USER_FAILED,
+      yield effects.put({
+        type: selectors$1.REGISTER_USER_FAILED,
         error: {
           message: 'Unable to parse the created user from the register service response'
         }
@@ -930,26 +956,26 @@ function* registerSaga({
     // Not OK responses, these can be due to service availability
     // or status codes echoed from the responses received from
     // management api when registering the user
-    const [, errorResponse] = yield to$1(response.json());
+    const [, errorResponse] = yield to.to(response.json());
     const error = errorResponse && errorResponse.error || errorResponse || {};
     // Get something meaningful from the response if there is no message in the body
     if (!error.message) {
       error.message = `Registration service: ${response.statusText}`;
       error.status = response.status;
     }
-    yield put({
-      type: REGISTER_USER_FAILED,
+    yield effects.put({
+      type: selectors$1.REGISTER_USER_FAILED,
       error
     });
   }
 }
 function* redirectSaga() {
   // Check if querystring contains a redirect_uri
-  const currentQs = queryParams(yield select(selectCurrentSearch));
+  const currentQs = selectors$1.queryParams(yield effects.select(selectors$1.selectCurrentSearch));
   const redirectUri = currentQs.redirect_uri || currentQs.redirect;
 
   // We must use redux based navigation to preserve the registration state
-  if (redirectUri) yield put(setRoute(redirectUri));
+  if (redirectUri) yield effects.put(selectors$1.setRoute(redirectUri));
 }
 
 const PAP_URL = 'https://pap.zengenti.com';
@@ -1017,80 +1043,80 @@ async function api(url, options) {
   });
 }
 
-const resetPasswordSagas = [takeEvery(REQUEST_USER_PASSWORD_RESET, requestPasswordResetSaga), takeEvery(RESET_USER_PASSWORD, resetPasswordSaga), takeEvery(CHANGE_USER_PASSWORD, changePasswordSaga)];
+const resetPasswordSagas = [effects.takeEvery(selectors$1.REQUEST_USER_PASSWORD_RESET, requestPasswordResetSaga), effects.takeEvery(selectors$1.RESET_USER_PASSWORD, resetPasswordSaga), effects.takeEvery(selectors$1.CHANGE_USER_PASSWORD, changePasswordSaga)];
 function* requestPasswordResetSaga(action) {
   const userEmailObject = action.userEmailObject;
-  yield put({
-    type: REQUEST_USER_PASSWORD_RESET_SENDING
+  yield effects.put({
+    type: selectors$1.REQUEST_USER_PASSWORD_RESET_SENDING
   });
   if (userEmailObject && userEmailObject.userEmail) {
     try {
       const passwordResetRequestResponse = yield UserHelper.RequestPasswordReset(userEmailObject);
       if (passwordResetRequestResponse) {
         if (!passwordResetRequestResponse.error) {
-          yield put({
-            type: REQUEST_USER_PASSWORD_RESET_SUCCESS
+          yield effects.put({
+            type: selectors$1.REQUEST_USER_PASSWORD_RESET_SUCCESS
           });
         } else {
-          yield put({
-            type: REQUEST_USER_PASSWORD_RESET_ERROR,
+          yield effects.put({
+            type: selectors$1.REQUEST_USER_PASSWORD_RESET_ERROR,
             error: passwordResetRequestResponse.error.message
           });
         }
       } else {
-        yield put({
-          type: REQUEST_USER_PASSWORD_RESET_ERROR,
+        yield effects.put({
+          type: selectors$1.REQUEST_USER_PASSWORD_RESET_ERROR,
           error: 'No response from server'
         });
       }
     } catch (error) {
-      yield put({
-        type: REQUEST_USER_PASSWORD_RESET_ERROR,
+      yield effects.put({
+        type: selectors$1.REQUEST_USER_PASSWORD_RESET_ERROR,
         error: error && error.toString()
       });
     }
   } else {
-    yield put({
-      type: REQUEST_USER_PASSWORD_RESET_ERROR,
+    yield effects.put({
+      type: selectors$1.REQUEST_USER_PASSWORD_RESET_ERROR,
       error: 'Invalid object'
     });
   }
 }
 function* resetPasswordSaga(action) {
   const resetPasswordObject = action.resetPasswordObject;
-  yield put({
-    type: RESET_USER_PASSWORD_SENDING
+  yield effects.put({
+    type: selectors$1.RESET_USER_PASSWORD_SENDING
   });
   if (resetPasswordObject.token && resetPasswordObject.password) {
     try {
       const resetPasswordResponse = yield UserHelper.ResetPassword(resetPasswordObject);
       if (resetPasswordResponse) {
         if (!resetPasswordResponse.error) {
-          yield put({
-            type: RESET_USER_PASSWORD_SUCCESS
+          yield effects.put({
+            type: selectors$1.RESET_USER_PASSWORD_SUCCESS
           });
         } else {
           const error = resetPasswordResponse.error.data && resetPasswordResponse.error.data.length > 0 && resetPasswordResponse.error.data[0].message || resetPasswordResponse.error.message;
-          yield put({
-            type: RESET_USER_PASSWORD_ERROR,
+          yield effects.put({
+            type: selectors$1.RESET_USER_PASSWORD_ERROR,
             error
           });
         }
       } else {
-        yield put({
-          type: RESET_USER_PASSWORD_ERROR,
+        yield effects.put({
+          type: selectors$1.RESET_USER_PASSWORD_ERROR,
           error: 'No response from server'
         });
       }
     } catch (error) {
-      yield put({
-        type: RESET_USER_PASSWORD_ERROR,
+      yield effects.put({
+        type: selectors$1.RESET_USER_PASSWORD_ERROR,
         error: error && error.toString()
       });
     }
   } else {
-    yield put({
-      type: RESET_USER_PASSWORD_ERROR,
+    yield effects.put({
+      type: selectors$1.RESET_USER_PASSWORD_ERROR,
       error: 'Invalid object'
     });
   }
@@ -1101,8 +1127,8 @@ function* resetPasswordSaga(action) {
 // newPassword
 function* changePasswordSaga(action) {
   if (!action || !action.userId || !action.currentPassword || !action.newPassword) {
-    yield put({
-      type: CHANGE_USER_PASSWORD_ERROR,
+    yield effects.put({
+      type: selectors$1.CHANGE_USER_PASSWORD_ERROR,
       error: 'Invalid action object sent to changePassword saga'
     });
     return;
@@ -1113,35 +1139,35 @@ function* changePasswordSaga(action) {
       existing: action.currentPassword,
       new: action.newPassword
     };
-    yield put({
-      type: CHANGE_USER_PASSWORD_SENDING
+    yield effects.put({
+      type: selectors$1.CHANGE_USER_PASSWORD_SENDING
     });
-    const clientCredentials = yield select(selectClientCredentials, 'js');
-    const client = yield getManagementApiClient({
+    const clientCredentials = yield effects.select(matchGroups.selectClientCredentials, 'js');
+    const client = yield ChangePassword_container.getManagementApiClient({
       ...clientCredentials
     });
-    const [err, res] = yield to$1(client.security.users.updatePassword(changePasswordObject));
+    const [err, res] = yield to.to(client.security.users.updatePassword(changePasswordObject));
     if (err) {
       var _err$data, _err$data2;
       const error = (err === null || err === void 0 || (_err$data = err.data) === null || _err$data === void 0 || (_err$data = _err$data.data) === null || _err$data === void 0 ? void 0 : _err$data.length) > 0 && err.data.data[0].message || (err === null || err === void 0 || (_err$data2 = err.data) === null || _err$data2 === void 0 ? void 0 : _err$data2.message);
-      yield put({
-        type: CHANGE_USER_PASSWORD_ERROR,
+      yield effects.put({
+        type: selectors$1.CHANGE_USER_PASSWORD_ERROR,
         error
       });
       return;
     }
-    yield put({
-      type: CHANGE_USER_PASSWORD_SUCCESS
+    yield effects.put({
+      type: selectors$1.CHANGE_USER_PASSWORD_SUCCESS
     });
   } catch (error) {
-    yield put({
-      type: CHANGE_USER_PASSWORD_ERROR,
+    yield effects.put({
+      type: selectors$1.CHANGE_USER_PASSWORD_ERROR,
       error: error && error.toString()
     });
   }
 }
 
-const userSagas = [...loginSagas, ...registerSagas, ...resetPasswordSagas];
+const userSagas = [...ChangePassword_container.loginSagas, ...registerSagas, ...resetPasswordSagas];
 
 var rootSaga = (featureSagas = []) => {
   return [...i18nSagas, ...routingSagas, ...navigationSagas, ...userSagas, ...featureSagas];
@@ -1194,8 +1220,13 @@ const pickProject = (hostname, query) => {
 };
 
 const AppRoot = props => {
-  return /*#__PURE__*/React.createElement(RouteLoader, props);
+  return /*#__PURE__*/React__default.default.createElement(RouteLoader.RouteLoader, props);
 };
 
-export { AppRoot as A, browserHistory as b, createLocaleRoutes as c, history as h, pickProject as p, rootSaga as r };
-//# sourceMappingURL=App-DvZXgw0e.js.map
+exports.AppRoot = AppRoot;
+exports.browserHistory = browserHistory;
+exports.createLocaleRoutes = createLocaleRoutes;
+exports.history = history;
+exports.pickProject = pickProject;
+exports.rootSaga = rootSaga;
+//# sourceMappingURL=App-Bo2cpotc.js.map
