+++ cjs/client.js
@@ -10,7 +10,7 @@ var reactRouterDom = require('react-router-dom');
 var component = require('@loadable/component');
 var queryString = require('query-string');
 var reactCookie = require('react-cookie');
-var App = require('./App-C2sEls2s.js');
+var App = require('./App-Bo2cpotc.js');
 var slice = require('./slice-5xJMH24n.js');
 var version = require('./version-CukCz8zL.js');
 var version$1 = require('./version-B9moFk8k.js');
@@ -33,7 +33,7 @@ require('./matchGroups-CxRa9Ej9.js');
 require('./CookieConstants-DfPiWCRZ.js');
 require('./CookieHelper.class-Det3qfdU.js');
 require('./ToJs-BsWqWjdm.js');
-require('./sagas-CmOxzqvB.js');
+require('./sagas-C6YSqau4.js');
 require('reselect');
 require('./util-CFAjDCA0.js');
 require('contensis-core-api');
