+++ cjs/contensis-react-base.js
@@ -8,7 +8,7 @@ var React = require('react');
 var reactRedux = require('react-redux');
 var slice = require('./slice-5xJMH24n.js');
 var mapJson = require('jsonpath-mapper');
-var sagas = require('./sagas-CmOxzqvB.js');
+var sagas = require('./sagas-C6YSqau4.js');
 require('reselect');
 require('immer');
 require('deep-equal');
@@ -32,7 +32,7 @@ var lodash = require('lodash');
 var lodashClean = require('lodash-clean');
 var CookieHelper_class = require('./CookieHelper.class-Det3qfdU.js');
 var cookiesMiddleware = require('universal-cookie-express');
-var App = require('./App-C2sEls2s.js');
+var App = require('./App-Bo2cpotc.js');
 var store = require('./store-DxuhGQ7p.js');
 var version = require('./version-B9moFk8k.js');
 var selectors = require('./selectors-C1CqEUmL.js');
