+++ cjs/search.js
@@ -3,7 +3,7 @@
 var React = require('react');
 var reactRedux = require('react-redux');
 var ToJs = require('./ToJs-BsWqWjdm.js');
-var sagas = require('./sagas-CmOxzqvB.js');
+var sagas = require('./sagas-C6YSqau4.js');
 var reselect = require('reselect');
 var immer = require('immer');
 var equals = require('deep-equal');
