+++ esm/client.js
@@ -6,8 +6,8 @@ import { unstable_HistoryRouter } from 'react-router-dom';
 import { loadableReady } from '@loadable/component';
 import { parse } from 'query-string';
 import { CookiesProvider } from 'react-cookie';
-import { c as createLocaleRoutes, b as browserHistory, r as rootSaga, p as pickProject } from './App-DvZXgw0e.js';
-export { A as ReactApp } from './App-DvZXgw0e.js';
+import { c as createLocaleRoutes, b as browserHistory, r as rootSaga, p as pickProject } from './App-Dy5DpY-p.js';
+export { A as ReactApp } from './App-Dy5DpY-p.js';
 import { a as actions } from './slice-C6JLQik8.js';
 import { s as selectVersionStatus } from './version-DlaBPQ7d.js';
 import { s as setVersionStatus } from './version-DIHIOZQD.js';
@@ -30,7 +30,7 @@ import './matchGroups-DT-RunAc.js';
 import './CookieConstants-DEmbwzYr.js';
 import './CookieHelper.class-C6rTRl_1.js';
 import './ToJs-BnRRHk6f.js';
-import './sagas-CFEhiMQT.js';
+import './sagas-CifdQEm-.js';
 import 'reselect';
 import './util-BuAZdfcq.js';
 import 'contensis-core-api';
