+++ esm/sagas-CifdQEm-.js
@@ -1292,10 +1292,10 @@ var expressions = /*#__PURE__*/Object.freeze({
   termExpressions: termExpressions
 });
 
-const filterQuery = (contentTypeIds, languages, versionStatus, customWhere) => {
+const filterQuery = (contentTypeIds, languages, versionStatus, customWhere, pageSize = 100) => {
   const query = new Query(...[...contentTypeIdExpression(contentTypeIds), ...languagesExpression(languages), ...defaultExpressions(versionStatus), ...customWhereExpressions(customWhere)]);
   query.orderBy = OrderBy.asc(Fields.entryTitle);
-  query.pageSize = 100;
+  query.pageSize = pageSize;
   return query;
 };
 const searchQuery = ({
@@ -5766,6 +5766,7 @@ function* loadFilter(action) {
   const {
     contentTypeId,
     customWhere,
+    pageSize,
     path
   } = filter;
   const createStateFrom = {
@@ -5781,7 +5782,7 @@ function* loadFilter(action) {
   try {
     if (contentTypeId) {
       const versionStatus = yield select(selectVersionStatus);
-      const query = filterQuery(Array.isArray(contentTypeId) ? contentTypeId : [contentTypeId], languages, versionStatus, customWhere);
+      const query = filterQuery(Array.isArray(contentTypeId) ? contentTypeId : [contentTypeId], languages, versionStatus, customWhere, pageSize);
       const payload = yield api.search(query, 0, projectId);
       if (!payload) throw new Error('No payload returned by search');
       if (payload.type === 'error') throw payload;
@@ -6041,4 +6042,4 @@ function* triggerSearchSsr(options) {
 }
 
 export { useFacets as $, updateCurrentTab$1 as A, updateCurrentFacet$1 as B, clearFilters$1 as C, selectListing as D, triggerSearch as E, Context as F, getFilters as G, UPDATE_SELECTED_FILTERS as H, UPDATE_SEARCH_TERM as I, UPDATE_PAGE_SIZE as J, UPDATE_PAGE_INDEX as K, SET_SEARCH_ENTRIES as L, SET_ROUTE_FILTERS as M, LOAD_FILTERS_COMPLETE as N, LOAD_FILTERS_ERROR as O, LOAD_FILTERS as P, EXECUTE_SEARCH_ERROR as Q, EXECUTE_SEARCH as R, SET_SEARCH_FILTERS as S, CLEAR_FILTERS as T, UPDATE_SORT_ORDER as U, APPLY_CONFIG as V, actions as W, selectors as X, types as Y, expressions as Z, queries as _, getTabsAndFacets$1 as a, useListing as a0, doSearch as a1, setRouteFilters as a2, searchSagas as a3, triggerListingSsr as a4, triggerMinilistSsr as a5, triggerSearchSsr as a6, defaultExpressions as a7, termExpressions as a8, contentTypeIdExpression as a9, filterExpressions as aa, orderByExpression as ab, customWhereExpressions as ac, cloneDeep as ad, getQueryParameter$2 as b, getSelectedFilters as c, getSearchTotalCount$1 as d, getSearchTerm$2 as e, getResultsInfo as f, getTotalCount$1 as g, getResults as h, getPageIsLoading$2 as i, getPaging as j, getIsLoading$2 as k, getRenderableFilters$2 as l, getFeaturedResults$2 as m, getFacetTitles$1 as n, getFacetsTotalCount$1 as o, getTabFacets$1 as p, getFacet$1 as q, getCurrentTab$1 as r, getPageIndex$2 as s, getCurrentFacet as t, updateSortOrder$1 as u, updateSelectedFilters as v, withMappers as w, updateSearchTerm$1 as x, updatePageSize$1 as y, updatePageIndex$1 as z };
-//# sourceMappingURL=sagas-CFEhiMQT.js.map
+//# sourceMappingURL=sagas-CifdQEm-.js.map
